Improve CanIUse data parsing
The caniuse-lite data for some supported features have extra notes after the `y` character (e.g. for the `:matches()` CSS pseudo-class the data for Safari 9+ reports `"y #2"`, referring to the note that _except_ the _full_ support of `:matches()` these browsers _also_ support the old prefixed version). Because of the strict comparison with `'y'`, these features were reported as non-supported, which doesn't seem correct to me.